requirements.txt
Django==4.2.9
djangorestframework==3.14.0
boto3==1.34.15
python-decouple==3.8